moved exerciserating related tests into package
